argsort variable names into plots.py to make available for other plot functions.
